+++ src/Event/SwitchDbEvent.php
@@ -2,6 +2,7 @@
 
 namespace Hakam\MultiTenancyBundle\Event;
 
+use Hakam\MultiTenancyBundle\Services\TenantDbConfigurationInterface;
 use Symfony\Contracts\EventDispatcher\Event;
 
 /**
@@ -9,15 +10,28 @@ use Symfony\Contracts\EventDispatcher\Event;
  */
 class SwitchDbEvent extends Event
 {
+    private ?TenantDbConfigurationInterface $tenant;
+
     private ?string $dbIndex;
 
-    public function __construct(?string $tenantDbIndex)
+    public function __construct(string|TenantDbConfigurationInterface|null $tenantOrTenantIdentifier)
     {
-        $this->dbIndex = $tenantDbIndex;
+        if ($tenantOrTenantIdentifier instanceof TenantDbConfigurationInterface) {
+            $this->tenant  = $tenantOrTenantIdentifier;
+            $this->dbIndex = $tenantOrTenantIdentifier->getIdentifierValue();
+        } else {
+            $this->tenant  = null;
+            $this->dbIndex = $tenantOrTenantIdentifier;
+        }
     }
 
     public function getDbIndex(): ?string
     {
         return $this->dbIndex;
+    }
+
+    public function getTenantDbConfiguration(): ?TenantDbConfigurationInterface
+    {
+        return $this->tenant;
     }
 }