+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v2.9.3.3
+* SwitchDbEvent – allow to set tenant entity
+* CreateDatabaseCommand – allow identifier to be string
+* DbSwitchEventListener - allow to use new SwitchDbEvent method to retrieve the tenant entity
+* introduced helper method TenantConnectionConfigDTO::fromTenantDbConfiguration - reduce code duplication
+
 # v2.9.3.2
 * update depencendy versions
 * convert services.xml to services.yaml